MORK::NDMORK
Struct NDMORK

pub struct NDMORK {
    pub stored_length: usize,
    pub error_fraction: f64,
    pub min_iter: u32,
    pub max_iter: u32,
    /* private fields */
}

NDMORK is an implement of node-determined multi-order Runge-Kutta methods.

Fields
stored_length: usizeerror_fraction: f64min_iter: u32max_iter: u32
Implementations

impl NDMORK

pub fn new( weights_function: Box<dyn Fn(u32) -> Vec<Vec<f64>>>, nodes: Vec<f64>, maximum_weight_graph: Vec<Vec<bool>>, ) -> Self

new creates a new instance of NDMORK. Requires the nodes, the weight function and the maximum weight digraph of the method.

pub fn set_minimum_length( h_powers: &mut Vec<f64>, weights_function: &dyn Fn(u32) -> Vec<Vec<f64>>, weights: &mut Vec<Vec<Vec<f64>>>, h: f64, s: usize, factorial: &mut Vec<f64>, stored_length: &mut usize, n: usize, implicit_ranks: &mut Vec<Vec<bool>>, computation_order: &Vec<SCC>, coefficients: &mut Vec<Vec<f64>>, nodes: &Vec<f64>, )

set_minimum_length ensures the method stores at least the constant necessary for initial value problems of order up to n