MORK
Module graph

Functions

SCC

SCC() is an implementation of Gabow’s algorithm, an algorithm which detects the strongly connected components of a digraph.

contraction

contraction takes a digraph and a partition of the nodes and outpus the contracted graph, without the loops.

priority

priority takes the cost of each node of a graph and outputs the priority of each node.

topological_sort

topological_sort takes a digraph with no non-empty walk and outputs a relabeling which topologically sorts the digraph.