MORK
Trait Solver

pub trait Solver {
    // Required method
    fn approximate(
        &mut self,
        t0: f64,
        h: f64,
        f: &dyn Fn(f64, &Vec<Vec<f64>>) -> Vec<f64>,
        y0: &Vec<Vec<f64>>,
    ) -> Vec<Vec<f64>>;
}

Solver is the used to indicate that a struct is a numerical scheme and can hence approximate the solution of an initial value problem.

Required Methods

fn approximate( &mut self, t0: f64, h: f64, f: &dyn Fn(f64, &Vec<Vec<f64>>) -> Vec<f64>, y0: &Vec<Vec<f64>>, ) -> Vec<Vec<f64>>

Given a differential equation function, initial instant, initial values, and a step size, approximate returns the approximation of the method/struct which implements this trait.